[iOS] Use result of main-frame Safe Browsing queries to cancel navigations
This CL makes SafeBrowsingTabHelper keep track of pending
main-frame queries issued during ShouldAllowRequest(),
and uses the results of these queries when producing the
decision for ShouldAllowResponse(). If the queried URL is
determined to be unsafe, the navigation is not allowed.
In the event that a Safe Browsing query has not completed
in time for ShouldAllowResponse(), the callback for this
method isn't invoked until a response to the query has
been received.

Bug: 1049340
Change-Id: Ib58246846bac37306247466aaf5049fd6ac21540
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173407
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#764766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0fa54135883c537315e12363f213630ba97947f1