[iOS] Get container from WebState delegate
This CL adds a method to get the container of the WebView from the
WebState delegate. It avoids using the keyWindow, which might be wrong
in multi window.
It also allows to use this delegate method to get the view used to know
if the Mobile or the Desktop UserAgent should be used.

Bug: 1073373
Change-Id: Ifeaa357441ec3fdb1b3fe3d5d4ac66ba22901319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172689
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765444}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e050a085eec05fa852cc9e8c513d5bfc4b89f5e5