ios: Clean up legacy webUI logic.
Don't restart or abort chrome URLs in didStartProvisionalNavigation.

The abort logic already exists in the
webview:decidePolicyForNavigationAction:decisionHandler callback.

The restart logic was there to recreate the webUI, but this can be done
within the same request.  Moving this logic to recreate the webUI to be
inside of WebStateImpl::CreateWebUI.

The bulk of the files changed are to store host in WebUIIOSController.

Bug: 546347
Change-Id: Ief66d6ca927b37f97c9359110d8144d841872350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179664
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765506}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe3ec66f698cb45ff4f8a7fb5a79764a6b38da66