[iOS] Present the long press bubble at startup
The current behaviour is that the long press bubble cannot be
triggered on iPad and is shown when opening a New Tab on iPhone.
This CL changes the behaviour to have it displayed when the BVC
becomes visible (i.e. at startup).

For now this experiment is disabled. As the current behaviour is weird,
it is better to have something consistent from which we can start
working if we want to re-launch this feature.

Fixed: 1077085
Change-Id: Ia171cd18a5d10c57ce1b5f37aebfc30bda1c4698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2178890
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765573}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18b01a92f6850e4423a5671e520ce132aec3a68c