[omnibox] Add prefs support for OmniboxPopupModel, add header test
This CL plumbs the PrefService through to OmniboxPopupModel, which is
needed to handle the collapsed header state.

It also starts making the collapsed headers properly supported, but
there's still more work to do.

It adds a test, but the test is DISABLED, since the implementation
to properly support collapsed headers is still incomplete.

That will be the very next CL.

Bug: 1078183, 1052522
Change-Id: I21b1f1689c6fcc904c6b9e3425f2737e0d4cf8d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183104
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766182}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db358e920b946a12e1e3ed0ed07b29d14573f6c5