[iOS] Fix testNoSearchForPDF on iPhone SE
The test was failing because the screen of the phone is too small to
display the action now that the "Downloads" entry is displayed.
Fix it be scrolling the menu.

Fixed: 1079185
Change-Id: I787bfeb9fdbcc183a247241d9f4cbe04489dd852
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187490
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#766372}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ae65597c4afb607146848851db88acec02cff17