[iOS] Always create Safe Browsing allow list and resource container.
Previously, the absence of these instances for a WebState was used as
a signal that a WebState was for prerender and would never need to
show a blocking page.  However, this makes the usage of these interfaces
tedious since they need to always include null checks.  Instead, this
CL updates these classes to always be created, and to check the
PrerenderService directly when choosing to use do-not-proceed without a
blocking page in UrlCheckerDelegateImpl.

Bug: 1064803
Change-Id: I59de04b04b74155f36bdafe3509edb898ec9a453
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183128
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767620}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee2bd1e1b872d476b03f991d9b85ba7e619d0986