[iOS] Add a WebStateObserver callback for redirects
When a navigation receives one or more redirects, Safe Browsing
needs to treat the navigation as unsafe if any URL in the
redirection chain is unsafe. It is not sufficient to only
consider the final URL in the chain (that is, the URL received
by ShouldAllowResponse). This means that SafeBrowsingTabHelper
needs to know about redirects, so that it can associate newer
queries with existing queries.

This CL adds a WebStateObserver callback for redirects. A
future CL will use this information in SafeBrowsingTabHelper.

Bug: 1049340
Change-Id: Ib4e0ad57f9bc5a167e72bf662448ab4b899f01d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194801
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768243}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4dd9f978baf5df9bb5e226cc1304b78b8bd8a07f