[ios] Create consistent primary/secondary button pointer interactions
Primary buttons (with blue backgrounds and white text) and secondary
buttons (with transparent backgrounds and blue text) are used in
several places in our app.

This CL creates consistent pointer interactions for these buttons.

By default, UIKit creates inconsistent hover effects for buttons
with opaque backgrounds depending on the placement and size of the
button. To create the most consistency, a custom HoverEffect is used
and the pointer shape does not change.

UIKit is also inconsistent for buttons with clear backgrounds. For these,
the size of the highlight pointer shape does not always match the size
of the button. To fix this, a custom pointer shape is set with the
size of the button.

This CL creates two util functions:
* CreateOpaqueButtonPointerStyleProvider() for buttons with an
opaque background.
* CreateTransparentButtonPointerStyleProvider() for buttons with a
transparent background.

As a rule of thumb:
* CreateOpaqueButtonPointerStyleProvider() should always be set for
primary buttons since UIKit's default behavior is very unpredictable.
* CreateTransparentButtonPointerStyleProvider() should be set for
secondary buttons, but especially long or large ones.
* However, most small clear background buttons do not need a custom
pointer style provider. Their pointer behavior is predictable.

This CL intentionally spans across multiple folders and owners
to demonstrate the usefulness of these utilities.

Test: iPad-only. Enable pointer support flag. Check consistency of pointer hover effects for primary and secondary buttons. Note: this does not affect third-party screens such as gaia.
Bug: 1073961
Change-Id: I93befd37e1dc916fc6f37901adee91913e0a7732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194803
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768324}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e271451e4422c65bf16d52dc15df914bbdc13299