[iOS] Support swipe to dismiss the sign-in consent
Adding support in UserSigninCoordinator for the swipe to dismiss
gesture. The swipe gesture is supported only if the sign-in flow is not
in progress. The gesture is blocked while the sign-in is in progress.

Fixed: 1081728
Change-Id: I78bd91291299eb1ecf6e3c1adc2fdd3ea3440aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199465
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#769211}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 39d62342dbe6e1c89aeab7185e103d8aef4b126e