[iOS] Add SafeBrowsingUrlAllowList::Observer
This observer class allows objects to observe changes in allow list
decision state.  Additionally, this CL introduces a Policy enum that
is used to describe the allow list decision state and updates the
implementation of the allow list to use this enum.  The allow list's
public interface will be updated in a follow-up CL to minimize the
number of files touched in this CL.

Bug: 1083557
Change-Id: I003d07dfc00c12eb473656e23b8b5157cda5cb57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205005
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#769880}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c3ee4c300d5859c4e5298e7895e82ce6a980a194