[ios] Allow the additional plist to be generated
Some of the additional plist needs to be generated based on some
build configuration, so add a new variable to control the list of
targets to add as dependency of the tweak_info_plist.

This variable will be overridden downstream with the list of
targets needed to generate the plist files.

Bug: 1083995
Change-Id: I73d9caad5e38e94c7c2fea43d6b59f97aec6e5dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207532
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#770088}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b94b4b17bb4bda405e34cfd0303426046c3a0a52