Add publicly_routable column to the history visits database
Indicates whether the IP of this URL was publicly routable. It’s a
property of the visit because it can change depending on the network
situation at the visit time. A value of “false” can mean several things:
the URL was not publicly routable; this was not a committed navigation;
this runs on iOS (unimplemented for iOS at this point); or the visit
was migrated.

The FloC project (I2P:
https://groups.google.com/a/chromium.org/g/blink-dev/c/DpZZG5K1PWs) will
need this bit, as it wants to query the history and only select
those URLs that had a publicly routable ip.


Change-Id: I499be8f9398ab885c71fca8d04d1b997171fcef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101323
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#770488}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ba9891006bbaba7e1899a984ab3be94c32b13cb