[iOS] Fix landscape popover presentation
Updating preferred ContentSize inside updateViewConstraints was causing
the popover arrow to not be pointing to the right place. Instead update
the preferred size when the trait collection changes and before the view
appears.

There is still a bit offset when the bar button is bigger than the
popover arrow and the button is on the screen margin. This is caused by
the popover being also layout at the margin without space to center the
arrow.

Bug: N/A
Change-Id: I10017b379c7688bea2cd6e514d0bf98a176c9c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215332
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772171}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a83543bf5fbe66eebfe69133c63662ee0d925691