Fix animation for landscape NTP
This CL fixes the display when showing the NTP on a cold start.
It prevents the animation by directly adding the BVC to the view
hierarchy when the animations are disabled.
It is also making sure that the location bar isn't shown on the NTP
when doing a cold start.

Fixed: 1082726
Change-Id: I02658837b177bbe07d58b507311657d09107fde0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216250
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 829c833a3c2e94f2203a13d8940b0e8dae873d6b