[iOS] Move HTTP auth completion callback to standalone C function.
This was previously handled by a WebStateDelegateTabHelper instance
method that was bound with a weak pointer.  If the WebState is destroyed
before the completion callback is executed, this means that the
WebKit-provided completion block will not be executed since the weak
pointer makes the callback a no-op.  This CL moves this logic into a
standalone C function that will execute the WebKit callback regardless
of whether the WebState has been destroyed.

Bug: 1087858
Change-Id: Id1a8ccfbead397d04b923b04e069589ea8e83967
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225461
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0dd7eee51ecbb766e829b1af6e54ed52a2803442