[iOS] Handle main-frame redirect chains when making Safe Browsing decisions
When a navigation produces a chain of server redirects, the
entire chain should be treated as unsafe if at least one
URL in the chain is unsafe.

This CL makes SafeBrowsingTabHelper keep track of the
redirect chain for the pending main frame query, and uses
the results of the queries for every URL in the redirect
chain when making a policy decision for the pending main
frame query.

This only handles the main frame case because there are no
WKWebView navigation callbacks informing the embedder about
sub-frame redirects.

Bug: 1049340
Change-Id: Ib7c296157558c95356546ad885833bb45a407e2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220274
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#774110}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a29caa72e1f7515165035cf2d50210e2b1744ee