[iOS] Disable flaky UKM EG test
testIncognitoPlusRegular fails flakily depending on
the order in which it is run relative to other test
cases. It also fails when it is the first test run
after app launch, which means that if it fails once
in a given run, it won't even pass on retry.

Here's an example run where this happens on a change
that is a no-op other than adding a new empty test:
https://chromium-swarm.appspot.com/task?id=4c998d1e469f0110

This CL disables this flaky test.

BUG=1033726

Change-Id: I77949094e448c26a5db176f4b18c60a783aefc8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231447
Commit-Queue: Ali Juma <ajuma@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Auto-Submit: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#775286}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 550c901a4a84a7c8b3663a2aa29434b9d252d582