[iOS] Simple mitigation for crash in -notifyForeground:
See the linked bug; under some circumstances, when the scene APIs are
*not* in use, [SceneController sceneState:transitionedToActivationLevel:]
is called with an SceneActivationLevelForegroundActive activation level
but without a the maininterface having a browser state (perhaps because
there is no main interface).

This simply bypasses the crashing call in this case (and the remainder
of the method is a no-op, since it all depends on the SceneState APIs
being used).

Bug: 1092326
Change-Id: Id0ab8427ba1d1a459345c2ebc20375ea19881bec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246678
Auto-Submit: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778358}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32b05b96ed5cd04adc5571466d9f40136575947a