[ios] Add ActivityToMoveTab to window_activities
This CL adds a new NSUserActivity type for moving a tab
to a new window.

This CL is one in a series to add drag and drop functionality
in a multi-window world.

Dragging a tab (which will come in a later CL) to the
edge of the screen in an iPad is a gesture to create
a new window with that tab. A custom NSUserActivity is
required to enable this gesture.

Bug: 1087844
Change-Id: I0462f3074825f682ccc4c1e8345f22e310abdcc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236896
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778848}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bea4cc9c332db4e893559aef2cf385262dfae76c