[iOS][MW] Prevent opening URL with FRE being presented
Fixing:
+[URLOpener openURL:applicationActive:tabOpener:startupInformation:]

This method needs to check if Chrome is presenting the FRE. With FRE
URL should be dropped.

This can be reproduced with multi-windows enabled, or when Chrome is
active when receiving the URL (this can be done with split-view).

Fixed: 1092492
Change-Id: Id9313e5eea484b262574d19c9ce30ceec1fab469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243132
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c916e991fe66895774703318bd7c1b91c13b2a9d