[ios] Add function to move tab across browsers
This CL adds a function MoveTab() to perform a tab move
from one Browser to another.

This CL is one in a series to add drag and drop functionality
in a multi-window world.

The algorithm looks for the source browser from which
to detach the tab. This code inefficiently loops through
a browser list to find the source browser. In practice,
this is not that bad since 1) this activity is not often
performed, and 2) the number of browsers in a browser list
is probably small.

Bug: 1087844
Change-Id: I2534e3232178fc1391156f862b0a9fabb0ec9004
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246783
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 033cb861bc2ca52691901d851cd6591b7be21579