[ios] Add drag and drop to tab grid
This CL brings drag and drop to tab grid.
- By default, the DragAndDrop feature flag is disabled
and the old gesture-based reordering is used.
- When the feature flag is enabled, reordering is done
via drag and drop.
- GridViewController fulfills the drag and drop interaction
delegate protocols since it is in the UI layer.
- The TabGridMediator creates the drag items (via helper utils)
and has the model logic to perform actions for the drop.
- A new protocol GridDragAndDropHandler serves as the liaison
between the UI and model layers.

This CL is one of several to add drag and drop functionality.

Bug: 1087844
Change-Id: I5add4a9192c300077b8f881e200a0090a1fce765
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245038
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779047}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eeafbbf4e5f1cf6fe9cec221ec08a0ff9ca99f40