[iOS] Fix bug around dismissing QR Code view after tapping Help
The fix is to get the root VC to dismiss our hierarchy. If we simply use
the current VC then, when the pop-over is presented, only that pop-over
will be dismissed (as per the UIKit documentation and as per this bug).

Bug: 1093908
Change-Id: Ia576910873acec1163f54abf8012210d13e016fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241995
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Auto-Submit: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779835}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 929d77a681920f5c2a6304a2331006ccea99c1d5