Re-enable ChromeWebClientTest, WKWebViewEarlyPageScriptAccessibility
The flaky part of this test is the same implementation as the other
tests in this file. The test is not flaky locally. Re-enable the test,
but also improve error logging to better debug this issue if it
re-occurs.

Fixed: 1098758
Change-Id: I9f09329f155a9b0824b88ed05cea4801cf6195b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264625
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#782202}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b54be785098cfde412eae2f176b5571ea1da0d4