[iOS] Fix the restoration of about: URL
This CL makes sure that when the navigation to about: URL finishes
the WebState is marked as no longer loading.
This is needed because the WKWebView is still marked as being loading.

This issue appeared with iOS 13.5 and seems fixed in iOS 14.

Fixed: 1099235, 1096900
Change-Id: I0ebfce4b832a439f740f8f372f47596572e91821
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2266553
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#782467}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 200bc6bb98a8daaf225d028a8ed0885b6e52e1d6