[ios] Use handler name instead of dispatcher for id<ApplicationCommands>
A 'dispatcher' is specifically an instance of CommandDispatcher, which
doesn't actually conform to protocols. For most non-coordinator classes,
the command dispatcher is an external implementation detail; all they
know is that they are passed some object that handles (hence the name)
commands as defined in some protocol.

Bug: None
Change-Id: Ife312df2f2dba1db346c740cac971f03868b1368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270686
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#783594}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc100446d1023bb3fb93b78e39bff585b3c43f83