[iOS] Cleanup TODO comments in UnifiedConsentCoordinator
-UnifiedConsentCoordinator.uiDisabled needs to remain.

crbug.com/1003737 was suggesting to make the SigninViewController
add ActivityOverlayCoordinator for the full sign-in view.
It would not be a good idea since the Cancel button would not be
tappable anymore.

Plus, an activity indicator is already displayed instead of "Yes I'm In"
button, when the sign-in process is triggered.

Fixed: 1003737
Change-Id: I57a5390b27ffa0f475a04962d7a53e3506224ac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273192
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#784026}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 687776b0c63bbcd8207069bbf004a51959cb6d83