[ios] Work around EarlGrey scrolling bugs.
EarlGrey recently added a new feature to wait for scroll bars to disappear after
a scroll operation. This is somehow causing subsequent taps to get interpreted
as a new scroll event, not a tap, which is leading to test failures.

As a workaround, this EarlGrey behavior can be disabled via a configuration
variable.

BUG=1101608

Change-Id: I2d4d9a928b057dcd6eedc1aebc6cb262b96882f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278763
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#784741}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecc19ffb7ff6e87cb5f07e28d74ca5a56b809bb2