[iOS] Add SceneStateBrowserAgent
This CL adds a browser agent (SceneStateBrowserAgent) to associate
a scene state with a Browser, so that coordinators and other objects
can access SceneState facilities if they have a Browser.

This approach is preferred over the (simpler) mechanism of just adding
a SceneState member to Browser. Conceptually that would be a layering
violation since SceneState is part of the UI layer.

BrowserViewWrangler is responsible for associating the SceneState with
each browser when it is created. Since this involves injecting UI-layer
information, it can't be handled in AttachBrowserAgents(). To do this,
BrowserViewWrangler is now initialized with a SceneState. This also
means that the BVW's sessionID doesn't need to be added separately.

Code that needs access to the scene state given a browser should use:

SceneStateBrowserAgent::FromBrowser(browser)->GetSceneState().

Bug: 1098268
Change-Id: I73997a848163e97d0ad579d862482b5a21fb1a71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278056
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785046}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c79522fbfbf722eb62b8a5904621ce5f563e7ef