[Enterprise iOS] Check savePasswordsItem before enabling the switch.
In the passwords settings page, the savePasswordsItem variable can be
unset when passwords are managed by policy. The
setSavePasswordsSwitchItemEnabled needs to check for that before
enabling/disabling it. Otherwise, it causes a nullptr crash when
swiping existing passwords to delete them.

Also fixes a bug introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/2219569
where the switch value would be reset to on when swiping a password.

Bug: 1101325
Change-Id: Icde314a7f1dccbee688f7fe882f44c7754c034b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281165
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: sebsg <sebsg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1089a553ecda7685e77409f03bc4f8adca263076