Safety Check iOS: Modify settings_password_check_cell for safety check
The custom cell for password check is similar to the one needed for
safety check.  It is being renamed to settings_check_cell for more
general use.

The main difference is the whether or not an image is used on the
leading side of the cell. This CL adds an optional image to the leading
side of the cell for compatibility with both checks.

Bug: 1078782
Change-Id: I825bdcb53f624ce2e567a68dc900bddc310dc270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273162
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785836}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f77680585a4d96f2c82f7ae4da71b4e9acc0ce6c