Execute external intents after scene is unblocked
If an external intent comes when Chrome shows a blocking UI, it should
be delayed until it is possible to execute it.

This CL considers the scene as not active if it is blocked.
The activities are stored in startupParameters when they are
received.
When the scene is unblocked, call handleExternalIntents to
execute it.

Bug: 1101786
Change-Id: I8117b0cd328847e03283f0d11fea825708594af5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292475
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788654}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70f823e553d872bf75ae8ecb7229d6fa8326e761