Do not create a second NTP on opening a window
New windows are created with an NSUserActivity to load
a URL (that can be chrome://newtabpage) in a new tab.

1. Use openOrReuseTab function to open such activities
to reuse the current tab if it is an NTP
2. avoid opening the default NTP if such an activity
exists as the NTP would flash before the target URL is
loaded.

Bug: 1106738
Change-Id: Ib034d54dc596b9ca5a69046fd71234831f35bd31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302616
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#789473}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7194d4944660d32623c0a5a8f195985de7f4726a