[iOS][MW] Fixing the sign-in upgrade promo
This fix is for Multi window. The sign-in upgrade promo has to be shown
in only one window when Chrome is started. This promo will be shown a
second time on a second start.

Instead of each window opening the sign-in upgrade promo, the main
controller needs to select only one window to show the promo:
-[MainController startUpAfterFirstWindowCreated] goes through the
list of scene to find the active one which can display the sign-in
upgrade promo.

Fixed: 1098188
Change-Id: I986b6b75e9f9a7c598381b68de766004e0259f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273281
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790507}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7c5eaf427ab630bf6b5dc6d7023f64fb0617af9